// src/utils/enums.ts
export enum ResourceType {
  WOOD = "Madera",
  STONE = "Piedra",
  FOOD = "Comida",
  GOLD = "Oro",
  SCIENCE = "Ciencia",
}

export enum TerrainType {
  PLAINS = "Llanura",
  FOREST = "Bosque",
  MOUNTAIN = "Montaña",
  WATER = "Agua",
  DESERT = "Desierto",
}

export enum UnitType {
  INFANTRY = "Infantería",
  CAVALRY = "Caballo",
  ELEPHANT = "Elefante",
  LEADER = "Líder",
  SHIP = "Barco",
}

export enum BuildingType {
  FORTRESS = "Fortaleza",
  TEMPLE = "Templo",
  MARKET = "Mercado",
  OBSERVATORY = "Observatorio",
  ACADEMY = "Academia",
  PORT = "Puerto",
  OBELISK = "Obelisco",
  SETTLEMENT = "Asentamiento",
}

export enum TechnologicalRootType {
  AGRICULTURE = "Agricultura",
  BUILDING = "Construcción",
  NAVIGATION = "Nautica",
  EDUCATION = "Educación",
  MILITARY = "Belicismo",
  RELIGION = "Espiritualidad",
  ECONOMY = "Economía",
  CULTURE = "Cultura",
  CIENCE = "Ciencia",
  DEMOCRACY = "Democracia",
  AUTOCRACY = "Autocracia",
  TEOCRACY = "Teocracia",
}

export enum TechnologicalSection {
  SECTION_1,
  SECTION_2,
  SECTION_3,
  SECTION_4,
}

export enum HappinessType {
  ANGRY = "angry",
  NEUTRAL = "neutral",
  HAPPY = "happy",
}

export enum IBOActions {
  ADVANCE = "Avanzar",
  RECRUIT = "Reclutar",
  ATACK = "Atacar",
  BUILD = "Construir",
  INFLUENCE = "Influir",
}